+++ migrations/install.php
@@ -17,11 +17,8 @@
         die('This program must be run from the command line.');
     }
 
-    // 1° Pre-flight check and bootup
-    // Check php version
-    if (version_compare(phpversion(), \UserFrosting\PHP_MIN_VERSION, "<")) {
-        die('UserFrosting requires PHP version '. \UserFrosting\PHP_MIN_VERSION.' or up.');
-    }
+    // Welcome message
+    echo "/****************************/\n/* UserFrosting's Migration */\n/****************************/";
 
     // Create new UserFrosting object, which will set up our DI container and boot up Sprinkles
     $uf = new UserFrosting();
@@ -43,7 +40,8 @@
     $dbParams = $config['db.default'];
 
     if (!$dbParams) {
-        die(PHP_EOL . "'default' database connection not found.  Please double-check your configuration.");
+        echo(PHP_EOL . "'default' database connection not found.  Please double-check your configuration.");
+        exit(1);
     }
 
     // Test database connection directly using PDO
@@ -53,7 +51,8 @@
         $message = PHP_EOL . "Could not connect to the database '{$dbParams['username']}@{$dbParams['host']}/{$dbParams['database']}'.  Please check your database configuration and/or google the exception shown below:" . PHP_EOL;
         $message .= "Exception: " . $e->getMessage() . PHP_EOL;
         $message .= "Trace: " . $e->getTraceAsString() . PHP_EOL;
-        die($message);
+        echo($message);
+        exit(1);
     }
 
     $schema = Capsule::schema();
@@ -152,4 +151,4 @@
     $uri = trim($uri, '/');
     */
 
-    echo "UserFrosting migrated successfully !".PHP_EOL;
+    echo "UserFrosting migrated successfully !".PHP_EOL.PHP_EOL;